requirements.txt
streamlit
fpdf
PyPDF2
Pillow